pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aikaboom"
version = "1.0.0"
description = "AIkaBoOM - AI Bills of Materials with conflict detection and SPDX 3.0.1"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "BOM Tools Contributors"}
]
keywords = ["bom", "ai", "dataset", "rag", "spdx", "metadata"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "flask>=3.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.20",
    "langchain-huggingface>=0.0.1",
    "langchain-text-splitters>=0.0.1",
    "langchain-core>=0.1.0",
    "langgraph>=0.0.20",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "PyGithub>=2.1.1",
    "huggingface-hub>=0.20.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pymupdf>=1.23.0",
    "certifi>=2023.0.0",
    "google-genai>=0.3.0",
    "httpx>=0.25.0",
    "pyshacl>=0.30.0",
    "jsonschema>=4.0.0",
]

[project.optional-dependencies]
link_fallback = [
    "google-genai>=0.3.0",
    "httpx>=0.25.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
aikaboom = "aikaboom.cli:main"

[project.urls]
Homepage = "https://github.com/rgopikrishnan91/aibom"
Documentation = "https://github.com/rgopikrishnan91/aibom/docs"
Repository = "https://github.com/rgopikrishnan91/aibom"
Issues = "https://github.com/rgopikrishnan91/aibom/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"aikaboom.web" = ["templates/*.html", "static/*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
